Refactor ZetkinUpdate type and TimelineUpdate component
