+++ src/locale/misc/updates/en.yml
@@ -1,3 +1,2 @@
-journeyInstance:
-  assignee:
-    add: '{actor} assigned {assignee}'
+journeyinstance:
+  addassignee: '{actor} assigned {assignee}'
