+++ src/types/zetkin.ts
@@ -277,15 +277,3 @@ export interface ZetkinJourneyMilestoneStatus {
   description: string;
   title: string;
 }
-
-export type ZetkinUpdateAssignee = {
-  assignee: Pick<ZetkinPerson, 'id' | 'first_name' | 'last_name'>;
-};
-
-export interface ZetkinUpdate {
-  actor?: Pick<ZetkinPerson, 'id' | 'first_name' | 'last_name'>;
-  created_at: string;
-  details: ZetkinUpdateAssignee;
-  type: 'journeyInstance.assignee.add';
-  target: { id: number; title: string };
-}
